// web/src/shared/api/generated/model/mathStepGradeDetailResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { MathStepScoreResult } from './mathStepScoreResult';

export interface MathStepGradeDetailResult {
  steps: MathStepScoreResult[];
  finalAnswerVerdict: string;
  justification: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  confidence: number | string;
  model: string;
  promptVersion: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  costUsd: number | string;
}